Add production readiness improvements: job alerting, CSRF, and fixes
Ops & Monitoring:
- Add job health alerting with email and Slack notifications
- Add POST /jobs/monitor-health endpoint for external cron
- Add alert deduplication (state transitions + 1hr cooldown)
- Add queue depth monitoring to /health endpoint

Security:
- Add CSRF protection on subscriber portal state-changing routes
- Validate Origin/Referer headers in production

Subscriber Portal:
- Use gross amounts (what subscriber paid) not net
- Add paymentCount to subscription list
- Add /subscriber/ to PUBLIC_ENDPOINTS (don't clear main auth on 401)

Onboarding Fixes:
- Allow decimal prices in singleAmount (was int-only)
- Add stepKey persistence for cross-device resume
- Add error fallback for username availability check
- Add tier normalization on currency change
- Add "Resuming..." text to loading shell

Tests:
- Add subscriber portal CSRF/amounts tests
- Add job health tracking and deduplication tests
- Add onboarding schema tests (decimal prices, stepKey)
- Update client.test.ts for subscriber portal auth isolation

+++ src/request/SelectRecipient.tsx
@@ -71,7 +71,8 @@ export default function SelectRecipient() {
         }
 
         setRecipient(recipient)
-        setAmount(parseInt(amount) || 0)
+        // Use parseFloat to preserve decimals (e.g., $10.50)
+        setAmount(parseFloat(amount) || 0)
         setIsRecurring(isRecurring)
         setPurpose(purpose)
 
@@ -84,7 +85,7 @@ export default function SelectRecipient() {
         navigate(-1)
     }
 
-    const canContinue = recipientName.trim().length > 0 && parseInt(amount) > 0
+    const canContinue = recipientName.trim().length > 0 && parseFloat(amount) > 0
 
     return (
         <div className="request-page request-page-modern">
